requirements.txt
torch
torchvision
torchaudio
torch-geometric
mediapipe
opencv-python
numpy
pyyaml
tqdm
